Record Long Number change history at full 64-bit fidelity (#871)
MetaverseObjectChangeAttributeValue had no long storage, so recording a
LongNumber attribute value into MVO change history narrowed it with an (int)
cast; any value beyond Int32 range (AD Large Integer attributes such as
accountExpires or pwdLastSet) was silently truncated in the audit record.

- Add LongValue (nullable bigint) to MetaverseObjectChangeAttributeValue with
  constructor, ToString rendering and an append-only EF migration; the
  AddAttributeValueChange LongNumber case now records the full value.
- Carry LongValue through the change history projection (MvoChangeHistoryDto,
  GetMvoChangeHistoryAsync) and its display string, which also serves the REST
  API change history endpoint and the portal timeline.
- Add LongValue and the missing DecimalValue to the bulk COPY column list in
  PersistPendingMvoChangesAsync: the Decimal work added the column and EF path
  but missed this raw SQL sibling, so Decimal change values written through
  the high-volume synchronisation path were silently dropped.
- New RequiresPostgres regression test proves both values survive the COPY
  path against real PostgreSQL (confirmed red before the fix, green after);
  in-memory tests structurally cannot see COPY column omissions. Model tests
  pin the full-fidelity recording, and the completeness guard test now drives
  LongNumber change history with a beyond-int-range value.

Full solution build clean (0 warnings); full test suite green (4,077 passed).

+++ src/JIM.Models/Core/MetaverseObjectChange.cs
@@ -154,10 +154,8 @@ public class MetaverseObjectChange
                     attributeChange, valueChangeType, (int)value.IntValue));
                 break;
             case AttributeDataType.LongNumber when value.LongValue != null:
-                // #871: MetaverseObjectChangeAttributeValue has no long storage; this narrows to int and can
-                // truncate values outside Int32 range. Tracked for an end-to-end Int64 audit.
                 attributeChange.ValueChanges.Add(new MetaverseObjectChangeAttributeValue(
-                    attributeChange, valueChangeType, (int)value.LongValue.Value));
+                    attributeChange, valueChangeType, value.LongValue.Value));
                 break;
             case AttributeDataType.Decimal when value.DecimalValue != null:
                 attributeChange.ValueChanges.Add(new MetaverseObjectChangeAttributeValue(
